laptop_service/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
django-environ==0.11.2
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
